fix: harden orders-service Kafka inbox and restore build

+++ orders-service/src/main/java/com/aszender/orders/dto/OrderItemResponse.java
@@ -1,9 +1,11 @@
 package com.aszender.orders.dto;
 
+import java.math.BigDecimal;
+
 public record OrderItemResponse(
         Long id,
         Long productId,
         Integer quantity,
-        Double unitPrice
+        BigDecimal unitPrice
 ) {
 }
